requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
azure-identity==1.24.0
azure-mgmt-compute==37.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-resourcehealth
azure-mgmt-network
azure-core==1.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.1
openai==1.99.9
agent-framework
lark-oapi>=1.4.14
